+++ test/convertors/ics-to-json-converter.ts
@@ -68,6 +68,26 @@ describe("IcsToJsonConverter", function () {
 
       expect(resultTyped).to.deep.equal(expectedResult);
     });
+
+    it("Recurring events - daylight saving time", async () => {
+      const expectedResult = await fs.readJson(path.join(__dirname, 'resources/recurring-events-dst.json'));
+      const ics = await fs.readFile(path.join(__dirname, 'resources/recurring-events-dst.ics'), 'utf-8');
+      const convertedRepresentation = await convertToJSON(ics);
+      const data = await readableToString(convertedRepresentation.data);
+      const resultTyped = JSON.parse(data);
+
+      expect(resultTyped).to.deep.equal(expectedResult);
+    });
+
+    it("Recurring events - daylight saving time 2", async () => {
+      const expectedResult = await fs.readJson(path.join(__dirname, 'resources/recurring-events-dst-2.json'));
+      const ics = await fs.readFile(path.join(__dirname, 'resources/recurring-events-dst-2.ics'), 'utf-8');
+      const convertedRepresentation = await convertToJSON(ics);
+      const data = await readableToString(convertedRepresentation.data);
+      const resultTyped = JSON.parse(data);
+
+      expect(resultTyped).to.deep.equal(expectedResult);
+    });
   });
 
   describe("Verify converter on incorrect input", () => {